requirements.txt
numpy==1.19.2
torchvision==0.9.1
torch==1.8.1
tqdm==4.50.2
opencv_python==4.5.1.48
albumentations==1.1.0
Pillow==9.1.1
segmentation_models_pytorch==0.2.1
ttach==0.0.3
